requirements.txt
streamlit>=1.24
tensorflow-cpu>=2.10
pillow
numpy
gdown
